add encoder.py
This change moves the transformer encoder layer and block under the
layers subpackage.
