+++ models/layers/__init__.py
@@ -4,3 +4,4 @@ from models.layers.patch_embed import PatchEmbedBlock, CeiTImage2TokenPatchEmbed
 from models.layers.position_embed import AddAbsPosEmbed
 from models.layers.attention import SelfAttentionBlock
 from models.layers.lca import LCAEncoder
+from models.layers.encoder import Encoder
